// apps/client/src/app/core/swagger/model/ngReportsEnvironment.ts
/**
 * NgReports
 * NG Reports API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { WindowSize } from './windowSize';

export interface NgReportsEnvironment { 
    userAgent: string;
    browserAppName: string;
    extensions: Array<Array<string>>;
    appEnvironment: string;
    appVersion: string;
    appLanguage: string;
    panelUrl: string;
    route: string;
    connectionType: string;
    ram: number;
    windowSize: WindowSize;
}